+++ src/intelligence_layer/core/__init__.py
@@ -1,6 +1,9 @@
 from .chunk import Chunk as Chunk
 from .chunk import ChunkInput as ChunkInput
 from .chunk import ChunkOutput as ChunkOutput
+from .chunk import ChunkWithIndices as ChunkWithIndices
+from .chunk import ChunkWithIndicesOutput as ChunkWithIndicesOutput
+from .chunk import ChunkWithStartIndex as ChunkWithStartIndex
 from .chunk import TextChunk as TextChunk
 from .detect_language import DetectLanguage as DetectLanguage
 from .detect_language import DetectLanguageInput as DetectLanguageInput